requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.3
daphne==4.0.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dotenv==1.1.0
razorpay==1.4.2
redis==6.0.0
requests==2.32.3
service-identity==24.2.0
setuptools==79.0.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
urllib3==2.4.0
zope.interface==7.2
